+++ requirements/pkg-extras.txt
@@ -6,6 +6,10 @@ sanic-cors
 fastapi >=0.63.0
 uvicorn[standard] >=0.13.4
 
+# extra=starlette
+fastapi >=0.16.0
+uvicorn[standard] >=0.13.4
+
 # extra=flask
 flask<2.0
 flask-cors
